Top Level Return Value Assertions
No assertions were being made on the return values of the top level class. This meant that swapping the return values of the top level class from a RESTClientResponse object to an integer saw the test still passing. In turn, making assertions on the return vales caused the rests to fail because the setUp fixture set out a contract to have no headers in the response. Closes #28